added new recipe w/@dcoopersmith
