--- TeachingKidsProgramming/src/org/teachingkidsprogramming/recipes/inDevelopment/PyramidsOfGizaStarter.java
@@ -1,39 +1,0 @@
-package org.teachingkidsprogramming.recipes.inDevelopment;
-
-import org.teachingextensions.logo.Tortoise;
-
-public class PyramidsOfGizaStarter
-{
-  /*----IN PROGRESS-----------------------------------*/
-  public static void main(String[] args) throws Exception
-  {
-    // Show the tortoise
-    Tortoise.show();
-    Tortoise.setSpeed(10);
-    Tortoise.turn(-90);
-    Tortoise.move(220);
-    Tortoise.turn(135);
-    Tortoise.move(100);
-    Tortoise.turn(90);
-    Tortoise.move(100);
-    Tortoise.turn(-90);
-    Tortoise.move(100);
-    Tortoise.turn(90);
-    Tortoise.move(100);
-    Tortoise.turn(-90);
-    Tortoise.move(100);
-    Tortoise.turn(90);
-    Tortoise.move(100);
-    Tortoise.turn(135);
-    Tortoise.move(210);
-  }
-}
-/*
-    //  Show the tortoise --#1
-    //  Make the tortoise move as fast as possible --#6
-    //  Do the following 4 times --#5.1
-    //      Change the pen color of the line the tortoise draws to blue --#4
-    //      Move the tortoise 50 pixels --#2
-    //      Turn the tortoise to the right (90 degrees) --#3
-    //  Repeat --#5.2
- */
