+++ TeachingKidsProgramming/src/org/teachingkidsprogramming/section02methods/PyramidsOfGizaEnd.java
@@ -1,0 +1,42 @@
+package org.teachingkidsprogramming.section02methods;
+
+import org.teachingextensions.logo.Tortoise;
+import org.teachingextensions.logo.utils.ColorUtils.PenColors;
+
+public class PyramidsOfGizaEnd
+{
+  public static void main(String[] args) throws Exception
+  {
+    setUpPyramidLand();
+    turnAndMove(-90, 220);
+    turnAndMove(135, 100);
+    moveDown();
+    moveUp();
+    moveDown();
+    moveUp();
+    moveDown();
+    turnAndMove(135, 210);
+  }
+  public static void turnAndMove(int degrees, int length)
+  {
+    Tortoise.turn(degrees);
+    Tortoise.move(length);
+  }
+  public static void moveDown()
+  {
+    turnAndMove(90, 100);
+  }
+  public static void moveUp()
+  {
+    turnAndMove(-90, 100);
+  }
+  private static void setUpPyramidLand()
+  {
+    Tortoise.show();
+    Tortoise.setSpeed(10);
+    Tortoise.getBackgroundWindow().setBackground(PenColors.Blues.AliceBlue);
+    Tortoise.setPenColor(PenColors.Yellows.DarkGoldenrod);
+    Tortoise.setPenWidth(2);
+    Tortoise.hide();
+  }
+}
