+++ TeachingKidsProgramming/src/org/teachingkidsprogramming/section02methods/PyramidsOfGizaMiddle.java
@@ -1,0 +1,38 @@
+package org.teachingkidsprogramming.section02methods;
+
+import org.teachingextensions.logo.Tortoise;
+// NOTE: Re-factor to remove duplicate code HINT: Use Extract Method
+import org.teachingextensions.logo.utils.ColorUtils.PenColors;
+
+public class PyramidsOfGizaMiddle
+{
+  public static void main(String[] args) throws Exception
+  {
+    setUpPyramidLand();
+    Tortoise.turn(-90);
+    Tortoise.move(220);
+    Tortoise.turn(135);
+    Tortoise.move(100);
+    Tortoise.turn(90);
+    Tortoise.move(100);
+    Tortoise.turn(-90);
+    Tortoise.move(100);
+    Tortoise.turn(90);
+    Tortoise.move(100);
+    Tortoise.turn(-90);
+    Tortoise.move(100);
+    Tortoise.turn(90);
+    Tortoise.move(100);
+    Tortoise.turn(135);
+    Tortoise.move(210);
+  }
+  private static void setUpPyramidLand()
+  {
+    Tortoise.show();
+    Tortoise.setSpeed(10);
+    Tortoise.getBackgroundWindow().setBackground(PenColors.Blues.AliceBlue);
+    Tortoise.setPenColor(PenColors.Yellows.DarkGoldenrod);
+    Tortoise.setPenWidth(2);
+    Tortoise.hide();
+  }
+}
